update Home UI

+++ src/net/xinzeling/setting/net/AppUpdateBean.java
@@ -6,14 +6,14 @@ import net.xinzeling.base.BaseResponseType;
 
 public class AppUpdateBean extends BaseResponseType{
 
-	public int version ;
+	public float version ;
 	public String downloadUrl ;
 	public String description ;
 	
 	@Override
 	public void parseJson(JSONObject json) {
 		try {
-			version = Integer.valueOf(json.optString("version"));
+			version = Float.valueOf(json.optString("version"));
 		} catch (Exception e) {
 		}
 		downloadUrl = json.optString("downloadUrl");
